requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0

# AI/ML frameworks
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20

# Enterprise integration
requests>=2.31.0
httpx>=0.25.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Logging and monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
